clean sample 04
