+++ 04-User-Profile/app/src/main/AndroidManifest.xml
@@ -14,13 +14,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
-        <activity android:name=".activities.StartActivity"
-            android:noHistory="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
 
         <!-- Auth0 Lock -->
         <activity
@@ -37,17 +30,23 @@
 
                 <data
                     android:host="@string/auth0_domain"
-                    android:pathPrefix="/android/YOURAPPPACKAGE/activities/callback"
+                    android:pathPrefix="/android/com.auth0.profiledemo/activities/callback"
                     android:scheme="https" />
             </intent-filter>
         </activity>
-        <!-- Auth0 Lock End -->
 
-        <activity android:name=".activities.LockActivity" />
-        <activity android:name=".activities.MainActivity"/>
         <activity
             android:name="com.auth0.android.provider.WebAuthActivity"
             android:theme="@style/Lock.Theme" />
+        <!-- Auth0 Lock End -->
+
+        <activity android:name=".activities.LoginActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity android:name=".activities.MainActivity" />
     </application>
 
 </manifest>
