--- 04-User-Profile/app/src/main/java/com/auth0/profiledemo/activities/StartActivity.java
@@ -1,23 +1,0 @@
-package com.auth0.profiledemo.activities;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import com.auth0.profiledemo.R;
-
-
-public class StartActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_start);
-
-        Intent lockIntent = new Intent(this, LockActivity.class);
-        startActivity(lockIntent);
-
-    }
-
-
-}
