+++ 04-User-Profile/app/src/main/java/com/auth0/profiledemo/application/App.java
@@ -4,21 +4,21 @@ import android.app.Application;
 
 import com.auth0.android.result.Credentials;
 
-public class App extends Application{
+public class App extends Application {
 
     private Credentials mUserCredentials;
 
     private static App appSingleton;
 
-    public static App getInstance(){
+    public static App getInstance() {
         return appSingleton;
     }
+
     @Override
     public void onCreate() {
         super.onCreate();
         appSingleton = this;
     }
-
 
     public Credentials getUserCredentials() {
         return mUserCredentials;
