+++ 04-User-Profile/app/src/main/res/layout/activity_main.xml
@@ -8,109 +8,101 @@
     android:padding="@dimen/activity_general_margin"
     tools:context=".activities.MainActivity">
 
-    <LinearLayout
-
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="30dp"
-        android:gravity="center"
-        android:orientation="vertical">
+        android:layout_height="wrap_content">
 
-        <ImageView
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:src="@drawable/logo"
-            tools:ignore="ContentDescription" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@android:color/white"
-        android:orientation="horizontal"
-        android:gravity="center_vertical"
-        android:padding="@dimen/activity_general_margin">
-
-        <ImageView
-            android:id="@+id/userPicture"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/activity_small_margin"
-            tools:ignore="ContentDescription" />
-
-        <RelativeLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
+            android:background="@android:color/white"
             android:gravity="center_vertical"
-            android:orientation="vertical">
+            android:orientation="vertical"
+            android:padding="@dimen/activity_general_margin">
 
-            <TextView
-                android:id="@+id/userNameTitle"
+            <ImageView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:src="@drawable/logo"
+                tools:ignore="ContentDescription" />
+
+            <ImageView
+                android:id="@+id/userPicture"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_alignParentLeft="true"
-                android:text="@string/username"
-                android:textColor="@android:color/black"
-                android:textSize="13sp" />
+                android:padding="@dimen/activity_small_margin"
+                tools:ignore="ContentDescription" />
 
-            <TextView
-                android:id="@+id/userEmailTitle"
+            <RelativeLayout
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_below="@id/userNameTitle"
-                android:layout_alignParentLeft="true"
-                android:text="@string/useremail"
-                android:textColor="@android:color/black"
-                android:textSize="13sp" />
+                android:gravity="center_vertical"
+                android:orientation="vertical">
 
-            <TextView
-                android:id="@+id/userCountryTitle"
-                android:layout_width="wrap_content"
+                <TextView
+                    android:id="@+id/userNameTitle"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_alignParentLeft="true"
+                    android:text="@string/username"
+                    android:textColor="@android:color/black"
+                    android:textSize="13sp" />
+
+                <TextView
+                    android:id="@+id/userEmailTitle"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_alignParentLeft="true"
+                    android:layout_below="@id/userNameTitle"
+                    android:text="@string/useremail"
+                    android:textColor="@android:color/black"
+                    android:textSize="13sp" />
+
+                <TextView
+                    android:id="@+id/userCountryTitle"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_alignParentLeft="true"
+                    android:layout_below="@id/userEmailTitle"
+                    android:text="@string/userCountry"
+                    android:textColor="@android:color/black"
+                    android:textSize="13sp"
+                    android:visibility="gone" />
+            </RelativeLayout>
+
+            <EditText
+                android:id="@+id/updateCountryEdittext"
+                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_below="@id/userEmailTitle"
-                android:layout_alignParentLeft="true"
-                android:text="@string/userCountry"
-                android:visibility="gone"
-                android:textColor="@android:color/black"
-                android:textSize="13sp" />
-
-        </RelativeLayout>
-
-    </LinearLayout>
-
-    <EditText
-        android:id="@+id/updateCountryEdittext"
-        android:layout_width="300dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:visibility="gone"
-        android:layout_margin="@dimen/activity_general_margin"
-        android:hint="@string/edit_hint"/>
+                android:layout_gravity="center"
+                android:layout_margin="@dimen/activity_general_margin"
+                android:hint="@string/edit_hint"
+                android:visibility="gone" />
 
 
-    <Button
-        android:id="@+id/editButton"
-        android:layout_width="300dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginTop="20dp"
-        android:background="@android:color/white"
-        android:enabled="true"
-        android:padding="@dimen/activity_general_margin"
-        android:text="@string/edit" />
+            <Button
+                android:id="@+id/editButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:layout_marginTop="20dp"
+                android:background="@android:color/white"
+                android:enabled="true"
+                android:padding="@dimen/activity_general_margin"
+                android:text="@string/edit" />
 
-    <Button
-        android:id="@+id/cancelEditionButton"
-        android:layout_width="300dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginTop="20dp"
-        android:background="@android:color/white"
-        android:padding="@dimen/activity_general_margin"
-        android:text="@string/cancel"
-        android:visibility="gone" />
-
+            <Button
+                android:id="@+id/cancelEditionButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:layout_marginTop="20dp"
+                android:background="@android:color/white"
+                android:padding="@dimen/activity_general_margin"
+                android:text="@string/cancel"
+                android:visibility="gone" />
+        </LinearLayout>
+    </ScrollView>
 
 
 </LinearLayout>
