+++ 04-User-Profile/app/src/main/res/values/strings.xml
@@ -1,18 +1,13 @@
 <resources>
     <string name="app_name">04-User Profile</string>
-
-    <string name="auth0_client_id">{CLIENT_ID}</string>
-    <string name="auth0_domain">{DOMAIN}</string>
-
     <string name="username">Name: %1$s</string>
     <string name="useremail">Email: %1$s</string>
     <string name="userCountry">Country: %1$s</string>
-
     <string name="edit_hint">Add the country…</string>
     <string name="edit">Edit Profile</string>
     <string name="save">Save Changes</string>
     <string name="cancel">Cancel Changes</string>
 
-
-
+    <string name="auth0_client_id">{CLIENT_ID}</string>
+    <string name="auth0_domain">{DOMAIN}</string>
 </resources>
